requirements.txt
numpy
Pillow
streamlit
Keras==2.4.3
tensorflow==2.3.1